fix: resolving PR comments
